BJS2-11800 - Mice are very nice!
+ fixed inaccuracies
